+++ R/function_MapRegionalSources.R
@@ -19,6 +19,7 @@
 #' @param fillOpacity Numeric, opacity of subbasin polygons in Leaflet maps. Used if \code{map} contains polygon data. See [leaflet::addPolygons()].
 #' @param line.weight Numeric, weight of connection lines in Leaflet maps. See [leaflet::addPolylines()].
 #' @param line.opacity Numeric, opacity of connection lines in Leaflet maps. See [leaflet::addPolylines()].
+#' @param seed Integer, seed number to to produce repeatable color palette.
 #' @param font.size Numeric, font size (px) for subbasin labels in Leaflet maps.
 #' @param file Save a Leaflet map to an image file by specifying the path to the desired output file using this argument. File extension must be specified.
 #' See [mapview::mapshot()].
@@ -63,7 +64,7 @@
 
 MapRegionalSources <- function(data, map, map.subid.column = 1, digits = 3, progbar = FALSE, map.type = "default",
                                plot.scale = TRUE, plot.searchbar = FALSE, weight = 0.5, opacity = 1, fillColor = "#4d4d4d",
-                               fillOpacity = 0.25, line.weight = 5, line.opacity = 1, font.size = 10, file = "",
+                               fillOpacity = 0.25, line.weight = 5, line.opacity = 1, seed = NULL, font.size = 10, file = "",
                                vwidth = 1424, vheight = 1000, html.name = "") {
 
   # Check/Load Dependencies for mapping features - do this here so that these packages are not required for the base HYPEtools installation
@@ -77,11 +78,10 @@ MapRegionalSources <- function(data, map, map.subid.column = 1, digits = 3, prog
     requireNamespace("leaflet", quietly = TRUE),
     requireNamespace("leaflet.extras", quietly = TRUE),
     requireNamespace("mapview", quietly = TRUE),
-    requireNamespace("htmlwidgets", quietly = TRUE),
-    requireNamespace("randomcoloR", quietly = TRUE)
+    requireNamespace("htmlwidgets", quietly = TRUE)
   )) {
     # Warn that a dependency is not installed
-    stop("To use this function, please ensure that the following packages are installed: sf, leaflet, leaflet.extras, mapview, htmlwidgets, randomcoloR", call. = FALSE)
+    stop("To use this function, please ensure that the following packages are installed: sf, leaflet, leaflet.extras, mapview, htmlwidgets", call. = FALSE)
 
     # Perform function
   } else {
@@ -231,9 +231,9 @@ MapRegionalSources <- function(data, map, map.subid.column = 1, digits = 3, prog
 
       # Create function to get colors for polylines
       color_pal <- function(X) {
-        tryCatch(randomcoloR::distinctColorPalette(X), # Try to get a distinct color for each line
+        tryCatch(distinctColorPalette(X, seed = seed), # Try to get a distinct color for each line
           error = function(e) {
-            rep_len(randomcoloR::distinctColorPalette(100), X) # If there is an error, then repeat palette of 100 colors as necessary
+            rep_len(distinctColorPalette(100, seed = seed), X) # If there is an error, then repeat palette of 100 colors as necessary
           }
         )
       }