+++ NEWS.md
@@ -14,6 +14,7 @@
 * Hide label group by default for Leaflet maps in MapRegionalSources()
 * Allow SUBIDs with many digits in SortGeoData()
 * Update PlotSimObsRegime() to ignore NA values and to fix percentile polygons when using log scale with negative/zero values (Issue #61)
+* Create distinctColorPalette() function and update PlotSubbasinRouting() and MapRegionalSources() to allow setting a seed so palettes can be reproduced
 * Various updates to help text
 
 ## HYPEtools 1.6.0 (2023-10-18)